--- feature/home/src/main/java/com/umc/home/test/HomeTestApplication.kt
@@ -1,10 +1,0 @@
-package com.umc.home.test
-
-import android.app.Application
-
-class HomeTestApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        // 테스트 환경을 위한 초기화 작업 추가 가능
-    }
-}
