+++ test/home/src/main/java/com/umc/ttatta/test/home/TestValues.kt
@@ -1,4 +1,4 @@
-package com.umc.home.test
+package com.umc.ttatta.test.home
 
 import java.time.LocalDateTime
 import com.umc.design.CategoryColor
@@ -9,8 +9,8 @@ object TestValues {
     const val NAME = "test_Na"
     const val NICKNAME = "Na"
     const val EMAIL = "test_Na@test.com"
-    val TODAY: LocalDateTime = LocalDateTime.parse("2025-02-01T15:15:30.000000")
-    const val CONTENT = "개빡치네"
+    val TODAY: LocalDateTime = LocalDateTime.parse("2025-02-04T15:15:30.000000")
+    const val CONTENT = "제발 잘 됐으면 좋겠다"
     const val LATITUDE = 37.5598451648105
     const val LONGITUDE = 126.920705864877
     const val CATEGORY_NAME = "test category"
